refactor to test plugin workflow
