+++ src/command/process_templates.rs
@@ -6,7 +6,7 @@ use crate::templates::{TemplateVariableProvider, DefaultTemplateVariableProvider
 use crate::templates::{TemplateConfigValidator, DefaultTemplateConfigValidator};
 use crate::templates::TemplateVariableReview;
 use crate::token_expander::{ExpandFilters, DefaultExpandFilters};
-use crate::plugin::{PluginRunner, DefaultPluginRunner};
+use crate::plugin::{PluginRunner, DefaultPluginRunner, PluginRunnerWorkflow};
 use crate::config::UserConfig;
 use std::format as s;
 
@@ -25,7 +25,7 @@ impl ProcessTemplates {
 
     // Runs any plugins that have been defined and updates template_variables with results
     let plugin_runner = DefaultPluginRunner::new();
-    plugin_runner.run_plugins(&mut template_variables)?;
+    PluginRunnerWorkflow::run_plugins(plugin_runner, &mut template_variables)?;
     VerboseLogger::log_template_variables(&user_config, &template_variables);
 
     // Ask for the user for the value of each variable
