+++ src/plugin/mod.rs
@@ -1,5 +1,9 @@
 mod plugin_runner;
 mod default_plugin_runner;
+mod plugin_result;
+mod plugin_runner_workflow;
 
 pub use default_plugin_runner::DefaultPluginRunner;
 pub use plugin_runner::PluginRunner;
+pub use plugin_result::PluginResult;
+pub use plugin_runner_workflow::PluginRunnerWorkflow;
